Moved helper functions into utils module.
