+++ src/lib/properties.js
@@ -5,68 +5,7 @@ const Layer = require('./properties/layer');
 const View = require('./properties/view');
 const TokendTransformer = require('./transformers/tokend');
 const Immutable = require('immutable');
-const isPlainObject = require('lodash.isplainobject');
-
-/* eslint-disable eqeqeq */
-/**
- * Deep-merge one Object into another. Do _not_ deep merge anything that isn't explicitly
- * a first-order instance of Object.
- *
- * @param  {Object} destination   The destination of the merge operation. This object is mutated
- * @param  {Object} source        The source that properties are merged from
- * @return {Object}               The destination object
- */
-const merge = (destination, source) => {
-  // Ensure that the destination value is an Object.
-  const dest = isPlainObject(destination) ? destination : {};
-
-  // Only merge source if it's an Object.
-  if (!isPlainObject(source)) {
-    return dest;
-  }
-
-  Object.keys(source).forEach((key) => {
-    // Ignore null and undefined source values. `== null` covers both
-    if (source[key] == null) {
-      return;
-    }
-
-    // Is this an Object (but not something that inherits Object)?
-    if (Object.getPrototypeOf(source[key]) === Object.prototype) {
-      // Recursively merge source Object into destination
-      dest[key] = merge(dest[key], source[key]);
-
-      return;
-    }
-
-    dest[key] = source[key];
-  });
-
-  return dest;
-};
-/* eslint-enable eqeqeq */
-
-/**
- * Recursively traverses a layer namespace and sets the value at the corresponding place in the object
- * @param {Object} destination The destination of the merge operation.
- * @param {Array<String>} namespaceArray An array of keys (namespaces) to traverse
- * @param {Object} source The source that properties are merged from
- * @return {Object}
- */
-const recusiveNamespaceMerge = (destination, namespaceArray, source) => {
-  const nextNamespace = namespaceArray.shift();
-  const dest = isPlainObject(destination) ? destination : {};
-
-  if (namespaceArray.length) {
-    dest[nextNamespace] = recusiveNamespaceMerge(dest[nextNamespace], namespaceArray, source);
-
-    return dest;
-  }
-
-  dest[nextNamespace] = merge(dest[nextNamespace], source);
-
-  return dest;
-};
+const Util = require('./util');
 
 /**
  * A Properties instance manages multiple statically configured layers,
@@ -187,24 +126,24 @@ class Properties extends EventEmitter {
       // template renderers.
       const persistent = this.layers.reduce((properties, layer) => {
         if (!layer.namespace) {
-          return merge(properties, layer.properties);
+          return Util.merge(properties, layer.properties);
         }
 
         let namespace = layer.namespace.split(':');
         const namespaceRoot = namespace.shift();
 
         if (namespace.length > 0) {
-          properties[namespaceRoot] = recusiveNamespaceMerge(properties[namespaceRoot], namespace, layer.properties);
+          properties[namespaceRoot] = Util.recusiveNamespaceMerge(properties[namespaceRoot], namespace, layer.properties);
         } else {
-          properties[namespaceRoot] = merge(properties[namespaceRoot], layer.properties);
+          properties[namespaceRoot] = Util.merge(properties[namespaceRoot], layer.properties);
         }
 
         return properties;
       }, {});
 
       this.persistent = persistent;
-      this._properties = merge(
-        this.active.sources.reduce((properties, source) => merge(properties, source.properties), {}),
+      this._properties = Util.merge(
+        this.active.sources.reduce((properties, source) => Util.merge(properties, source.properties), {}),
         persistent
       );
 
@@ -221,6 +160,6 @@ Properties.BUILD_HOLD_DOWN = 1000; // eslint-disable-line rapid7/static-magic-nu
 
 Properties.Layer = Layer;
 Properties.View = View;
-Properties.merge = merge;
+Properties.merge = Util.merge;
 
 module.exports = Properties;
